refactor: refactor constraint
